requirements.txt
bidict==0.22.0
click==8.1.2
colorama==0.4.4
Flask==2.1.1
Flask-SocketIO==5.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
MouseInfo==0.1.3
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-engineio==4.3.1
python-socketio==5.5.2
pytweening==1.0.4
Werkzeug==2.1.1
